recadtor: update package structure
