+++ .eslintrc.yml
@@ -1,0 +1,1 @@
+extends: 'eslint-config-cheminfo'
