+++ .gitignore
@@ -1,2 +1,6 @@
-.idea
 node_modules
+.idea
+logs
+*.log
+npm-debug.log*
+coverage
