+++ .travis.yml
@@ -1,4 +1,7 @@
 language: node_js
 node_js:
-  - "0.12"
-  - "iojs"
+  - node
+  - lts/boron # v6
+  - lts/argon # v4
+script: "npm run test-travis "
+after_script: "bash <(curl -s https://codecov.io/bash)"
