+++ .yo-rc.json
@@ -1,0 +1,8 @@
+{
+  "generator-cheminfo": {
+    "promptValues": {
+      "userName": "Miguel Asencio",
+      "email": "maasencioh@gmail.com"
+    }
+  }
+}
