+++ package.json
@@ -2,41 +2,47 @@
   "name": "ml-preprocess",
   "version": "1.0.0",
   "description": "preprocess functions to improve the representation and quality of data",
-  "main": "src/index.js",
-  "directories": {
-    "lib": "src",
-    "test": "test"
-  },
+  "main": "./src/index.js",
+  "files": [
+    "runkit.js",
+    "src"
+  ],
   "scripts": {
-    "test": "mocha --require should-approximately-deep --reporter mocha-better-spec-reporter --recursive"
+    "eslint": "eslint src test",
+    "eslint-fix": "npm run eslint -- --fix",
+    "test": "npm run test-mocha && npm run eslint",
+    "test-cov": "istanbul cover node_modules/.bin/_mocha -- --require should --reporter dot --recursive",
+    "test-travis": "istanbul cover node_modules/.bin/_mocha --report lcovonly -- --require should --reporter dot --recursive",
+    "test-mocha": "mocha --require should --reporter mocha-better-spec-reporter"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/mljs/preprocess.git"
   },
   "keywords": [
-    "preprocess",
-    "metric",
-    "vector",
+    "machine",
+    "learning",
     "data",
     "mining",
-    "datamining",
-    "machine",
-    "learning"
+    "datamining"
   ],
-  "author": "Miguel Asencio",
+  "author": "Miguel Asencio <maasencioh@gmail.com> (https://github.com/Miguel Asencio)",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/mljs/preprocess/issues"
   },
-  "homepage": "https://github.com/mljs/preprocess",
+  "runkitExampleFilename": "./runkit.js",
+  "homepage": "https://github.com/mljs/preprocess#readme",
   "devDependencies": {
-    "mocha-better-spec-reporter": "latest",
-    "mocha": "latest",
-    "should": "latest",
-    "should-approximately-deep": "latest"
+    "eslint": "^3.12.2",
+    "eslint-config-cheminfo": "^1.6.0",
+    "eslint-plugin-no-only-tests": "^1.1.0",
+    "istanbul": "^0.4.5",
+    "mocha": "^3.2.0",
+    "mocha-better-spec-reporter": "^3.1.0",
+    "should": "^11.1.2"
   },
   "dependencies": {
-    "ml-stat": "^1.0.1"
+    "ml-stat": "^1.3.3"
   }
 }
